+++ src/modules/state.ts
@@ -474,14 +474,8 @@ function _declareState(scope) {
 		PostAdReentryContinuationMs:
 			_C.POST_AD_REENTRY_CONTINUATION_MS ?? 15000,
 		AdEndGraceMs: _C.AD_END_GRACE_MS ?? 2500,
-		AdEndMaxWaitMs: _C.AD_END_MAX_WAIT_MS ?? 2500,
 		AdEndMinCleanPlaylists: _C.AD_END_MIN_CLEAN_PLAYLISTS ?? 2,
-		AdEndMinNativeRecoveryProbes:
-			_C.AD_END_MIN_NATIVE_RECOVERY_PROBES ?? 3,
-		AdEndNativeRecoveryProbeCooldownMs:
-			_C.AD_END_NATIVE_RECOVERY_PROBE_COOLDOWN_MS ?? 750,
 		AdRecoveryReloadCooldownMs: _C.AD_RECOVERY_RELOAD_COOLDOWN_MS ?? 10000,
-		CleanBackupReuseMs: _C.CLEAN_BACKUP_REUSE_MS ?? 1500,
 		HasTriggeredPlayerReload: false,
 		PendingTriggeredPlayerReloadChannel: null,
 		PendingTriggeredPlayerReloadMediaKey: null,